+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Changelog
-### 0.0.9 WIP
-    * inheriting binding an queue names
+
+### 0.1.0 
+    * consumer server
+### 0.0.9
+    * inheriting binding and queue names
+    * hop_hop binary (running consumers)
 ### 0.0.8
     * return state os consume documented and changed (true for exit_loop, false for exceptional exit)